refactor: StopAudioButton 위치 수정, 음성 파일 추가
